moved Ruben's Admin files into api path
